+++ Changelog.md
@@ -15,7 +15,7 @@ v2.11.0 (not published yet)
 - Fix remove confirmation on Lists showing first item name only even if multiple items are selected
 - Reorganize encoder profiles
 - Add NVEnc parameters --lut3d, --lut3d_interp, --dolby-vision-rpu, --dolby-vision-profile
-- Add VCEEnc parameter --thread-affinity
+- Add VCEEnc parameter --thread-affinity, --qvbr, --qvbr-quality
 - Add QSVEnc parameters --thread-affinity, --dhdr10-info, --dolby-vision-rpu, --dolby-vision-profile
 - Add x265 Dolby Vision Profile 8.4
 - Fix minor mkvextract demux issue ([#833](/../../issues/833))
@@ -25,6 +25,7 @@ v2.11.0 (not published yet)
 - Optimize Check for Updates
 - Show also the path of the project when it's not found
 - Add "Demux Attachments" to project options
+- Add "Bitrate" to VCEEnc and QSVEnc options
 - Update tools
   - MKVToolNix v66.0
   - NVEnc v5.46